third_party/blink/web_tests/external/wpt/workers/interfaces/WorkerGlobalScope/onerror/message-classic-DOMException-expected.txt

This is a testharness.js-based test.
FAIL Throw DOMException-TypeError in toplevel: classic: listener assert_regexp_match: e.message should contain the message of the thrown error expected object "/Throw in/" but got "Uncaught "
FAIL Throw DOMException-TypeError in toplevel: classic: handler assert_regexp_match: e.message should contain the message of the thrown error expected object "/Throw in/" but got "Uncaught "
PASS Throw DOMException-TypeError in setTimeout-function: classic: listener
PASS Throw DOMException-TypeError in setTimeout-function: classic: handler
FAIL Throw DOMException-TypeError in setTimeout-string: classic: listener assert_regexp_match: e.message should contain the message of the thrown error expected object "/Throw in/" but got "Uncaught "
FAIL Throw DOMException-TypeError in setTimeout-string: classic: handler assert_regexp_match: e.message should contain the message of the thrown error expected object "/Throw in/" but got "Uncaught "
PASS Throw DOMException-TypeError in onmessage: classic: listener
PASS Throw DOMException-TypeError in onmessage: classic: handler
PASS Throw DOMException-TypeError in onerror: classic: listener
PASS Throw DOMException-TypeError in onerror: classic: handler
Harness: the test ran to completion.
